Remove redundant casts of argument to (prototyped) double parameter.

+++ include/land.h
@@ -142,27 +142,27 @@ struct lchrstr {
 #define L_TRAIN         bit(11)	/* train unit - neato */
 #define L_HEAVY         bit(12)	/* heavy unit - can't go on trains */
 
-#define LND_ATTDEF(b, t) (((b) * (1.0 + ((sqrt((double)(t)) / 100.0) * 4.0))) \
-						  > 127 ? 127 : \
-						  ((b) * (1.0 + ((sqrt((double)(t)) / 100.0) * 4.0))))
-#define LND_SPD(b, t) ((b * (1.0 + ((sqrt((double)t) / 100.0) * 2.1))) > 127\
-		       ? 127 : (b * (1.0 + ((sqrt((double)t) / 100.0) * 2.1))))
-#define LND_VUL(b, t) ((b * (1.0 - ((sqrt((double)t) / 100.0) * 1.1))) < 0\
-		       ? 0 : (b * (1.0 - ((sqrt((double)t) / 100.0) * 1.1))))
+#define LND_ATTDEF(b, t) (((b) * (1.0 + ((sqrt((t)) / 100.0) * 4.0)))	\
+			  > 127 ? 127 :					\
+			  ((b) * (1.0 + ((sqrt((t)) / 100.0) * 4.0))))
+#define LND_SPD(b, t) ((b * (1.0 + ((sqrt(t) / 100.0) * 2.1))) > 127	\
+		       ? 127 : (b * (1.0 + ((sqrt(t) / 100.0) * 2.1))))
+#define LND_VUL(b, t) ((b * (1.0 - ((sqrt(t) / 100.0) * 1.1))) < 0	\
+		       ? 0 : (b * (1.0 - ((sqrt(t) / 100.0) * 1.1))))
 #define LND_VIS(b, t) (b)
 #define LND_SPY(b, t) (b)
 #define LND_RAD(b, t) (b)
-#define LND_FRG(b, t) ((t) ? \
-					   ((b) * (logx((double)(t), (double)35.0) < 1.0 ? 1.0 : \
-							   logx((double)(t), (double)35.0))) : (b))
-#define LND_DAM(b, t) ((t) ? \
-					   ((b) * (logx((double)(t), (double)60.0) < 1.0 ? 1.0 : \
-							   logx((double)(t), (double)60.0))) : (b))
-#define LND_ACC(b, t) ((b * (1.0 - ((sqrt((double)t) / 100.0) * 1.1))) < 0\
-		       ? 0 : (b * (1.0 - ((sqrt((double)t) / 100.0) * 1.1))))
+#define LND_FRG(b, t) ((t) ?				     \
+		       ((b) * (logx((t), 35.0) < 1.0 ? 1.0 : \
+			       logx((t), 35.0))) : (b))
+#define LND_DAM(b, t) ((t) ?				     \
+		       ((b) * (logx((t), 60.0) < 1.0 ? 1.0 : \
+			       logx((t), 60.0))) : (b))
+#define LND_ACC(b, t) ((b * (1.0 - ((sqrt(t) / 100.0) * 1.1))) < 0	\
+		       ? 0 : (b * (1.0 - ((sqrt(t) / 100.0) * 1.1))))
 #define LND_AMM(b, d, t) ((b) ? ((LND_DAM((d), (t)) / 2) + 1) : 0)
-#define LND_AAF(b, t) ((b * (1.0 + ((sqrt((double)t) / 100.0) * 3.0))) > 127\
-		       ? 127 : (b * (1.0 + ((sqrt((double)t) / 100.0) * 3.0))))
+#define LND_AAF(b, t) ((b * (1.0 + ((sqrt(t) / 100.0) * 3.0))) > 127	\
+		       ? 127 : (b * (1.0 + ((sqrt(t) / 100.0) * 3.0))))
 #define LND_FC(b, t)  (b)
 #define LND_FU(b, t)  (b)
 #define LND_XPL(b, t) (b)
